requirements.txt
requests
pynacl
python-dotenv